Move fix-point analyses to their own module
