+++ src/ir/context.rs
@@ -5,12 +5,10 @@ use super::int::IntKind;
 use super::item::{IsOpaque, Item, ItemAncestors, ItemCanonicalPath, ItemSet};
 use super::item_kind::ItemKind;
 use super::module::{Module, ModuleKind};
-use super::named::UsedTemplateParameters;
-use super::analysis::analyze;
+use super::analysis::{analyze, UsedTemplateParameters, CantDeriveDebugAnalysis};
 use super::template::{TemplateInstantiation, TemplateParameters};
 use super::traversal::{self, Edge, ItemTraversal};
 use super::ty::{FloatKind, Type, TypeKind};
-use super::cant_derive_debug::CantDeriveDebugAnalysis;
 use BindgenOptions;
 use cexpr;
 use callbacks::ParseCallbacks;
