+++ src/ir/mod.rs
@@ -17,10 +17,8 @@ pub mod item;
 pub mod item_kind;
 pub mod layout;
 pub mod module;
-pub mod named;
 pub mod template;
 pub mod traversal;
 pub mod ty;
 pub mod var;
 pub mod objc;
-pub mod cant_derive_debug;
